[lint-unknown-licenses] Reformatted code, allow small typos
